debugged
pay extra attention to the length of the logic vector when perform bit
shifting. also, fixed the bug in search. when we are sure the allocation
can be decide in the current depth,  find the first OR = 0. (when we
need to go further down the tree, we find the first AND = 0)